IMPALA-11807: Rewrite iceberg metadata if not on hdfs
Iceberg test tables are usually written on hdfs and the file paths start
with "hdfs://localhost:20500/test-warehouse".

Earlier we manually transformed the metadata so paths would start with
"/test-warehouse"

Since IMPALA-11821, testdata/bin/rewrite-iceberg-metadata.py supports
not only a custom WAREHOUSE_LOCATION_PREFIX, but the ability to trim the
beginning of the file paths.

This commit modifies the data load, so metadata rewrite always executes
if not on hdfs, even with empty WAREHOUSE_LOCATION_PREFIX.

Testing:
  - Ran iceberg tests on ozone and S3

Change-Id: Ic04c5abdd42cb0c1cf5abd310b06c39cf8cd64ba
Reviewed-on: http://gerrit.cloudera.org:8080/19432
Reviewed-by: Michael Smith <michael.smith@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>