IMPALA-11756: Disable auto analyze table triggered by Hive
Hive will automatically recompute stats asynchronously after major
compactions. The tasks could fail if the table is removed, which usually
happens on our test tables. The failure could cause further failures on
other query-based compactions in the same session, and results in test
failures. See the analysis in the JIRA for an example.

We don't rely on Hive to recompute the stats. So we can disable this
feature to avoid the issue. This patch turns off
"hive.compactor.gather.stats" to disable it.

Tests:
 - Ran exhaustive tests

Change-Id: Idc23100ae74d6cb07894053a26806e01258065ec
Reviewed-on: http://gerrit.cloudera.org:8080/19464
Reviewed-by: Riza Suminto <riza.suminto@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>