IMPALA-11899: Add mandatory links to the Impala Apache home page
The Apache Foundation has a number of requirements for TLP websites,
which are checked regularly, with reports being published
whimsy.apache.org.

Impala failed in two categories: missing a copyright notice, and missing
a link to the Apache Privacy Policy.

This change to the Impala home page adds both.

It also fixes a misindented source line, and wraps the whole footer into
the <center> tag originally reserved for the events box.

Tested by loading the local copy of the page into Chrome and Firefox on
my system.

Change-Id: Ibbb491fe2ec3c15305c3c66f1e8857a98fced10f
Reviewed-on: http://gerrit.cloudera.org:8080/19477
Reviewed-by: Michael Smith <michael.smith@cloudera.com>
Tested-by: Laszlo Gaal <laszlo.gaal@cloudera.com>