IMPALA-11869: Use to_string() in PrintValue for printing Thrift types
IMPALA-11645 introduced the function PrintValue() which we use to
convert Thrift types to strings. This function used operator<<,
which is provided for the generated Thrift types. However, Thrift also
generates a to_string() function that is overloaded for Thrift types. It
would be more efficient to use this instead of involving streams with
operator<<.

This change reimplements PrintValue() using to_string() instead of
operator<<.

Change-Id: Ibc5b847dea2bdea7ba0ab8e093a8bab5a8145019
Reviewed-on: http://gerrit.cloudera.org:8080/19487
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>