IMPALA-11916: Replace base::IsAarch64 with constant
base::IsAarch64() returns true for Aarch64 platforms and returns false
for x86 platforms. It shows up in a perf test and consumes some
percents. Note that it's used in the hot path of hash computation. This
patch replaces this method with a constant, IS_AARCH64. Defines the
constant in cpu-info.h since the original place is in a file ported from
gutil.

Test:
 - Redo the perf test and don't see base::IsAarch64 anymore.

Change-Id: Id6d62de63a0cb7b94244a1d4f8dcc6b2e65b6d9c
Reviewed-on: http://gerrit.cloudera.org:8080/19492
Reviewed-by: Wenzhe Zhou <wzhou@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>