IMPALA-11925: Added a new api add_write_notification_log_in_batch in the MetastoreServiceHandler class
In the MetastoreServiceHandler class, the 'add_write_notification_log_
in_batch' API implementation is missing which is causing test failures.
This needs to be fixed by adding implementation of the 'add_write_
notification_log_in_batch' api in the MetastoreServiceHandler class

Testing:
Added a test in the 'test_hms_apis()' to verify that the 'add_write_
notification_log_in_batch' api in HMS is reachable from impala. The
same test can be used in the future to test newly added apis in HMS.

Change-Id: Ia9a35557c2ed79ed0276c4a418b5292fc6dd6194
Reviewed-on: http://gerrit.cloudera.org:8080/19508
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>