IMPALA-10111: Fix TestWebPage::test_query_stmt flakiness
A query might be logged in "in_flight_queries" even though it has
already finished. Earlier the test only looked for the executed query
in "completed_queries", after this commit it will look in both lists.

Moreover, removed unnecessary duplicated execution of the query and
added formatting to dumped response to ease debugging in the future.

Testing:
 - Locally reproducible for the first test run after starting the
cluster.
 - Verified that the query is found in in_flight_queries too and the
test passes in this case.

Change-Id: If25d430a871415a3884dece772b8d8105c583c05
Reviewed-on: http://gerrit.cloudera.org:8080/19524
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>