IMPALA-11823: Add more items to Impala web UI queries page
When operating and maintaining an Impala cluster or analyzing historical
query performance, it will be helpful if we show the memory consumed,
the amount of data read, and other information of the query from the
historical query page of the web UI. The current historical query page
does not display these information, so we should count this information
when the query is executed and display it on the web page.

This patch modifies the query list page (/queries) and query detail
pages (/query_plan, etc.).
On the list page, some metrics are added for each query record,
including queuing time, memory usage, memory estimation, bytes read, and
bytes sent. In addition, the Details column now shows the query ID and
the position is adjusted to make them at the top of the record for easy
clicking.
On the query detail page, a similar record table is added to display the
key information of the current query. In addition, a timeline display is
added to the summary page (which is exactly the same as the timeline in
the profile, just for quick viewing). For queries that are running, the
above information will be automatically refreshed (only for the plan and
summary tabs).
To make it clear what each metric means, tooltips are added to all list
headers.

Change-Id: I19c75461a6405025fa433ae84d2c94d013fcaacb
Reviewed-on: http://gerrit.cloudera.org:8080/19417
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>