IMPALA-11950: Planner change for Iceberg metadata querying
This commit extends the planner with IcebergMetadataScanNode which will
be used to scan Iceberg metadata tables (IMPALA-10947). The scan node is
only implemented on the frontend side in this patch, the backend part
will be developed in IMPALA-11996.

To avoid executing the plan there is a hardcoded condition, it is after
the explain part, so the change remains testable with EXPLAIN queries.

Testing:
 - Added planner tests

Change-Id: I3675d7a57ca570bfec306798589b5ef6aa34b5c6
Reviewed-on: http://gerrit.cloudera.org:8080/19547
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>