IMPALA-11795: Ignore high/low values stats for timestamp columns
Timestamp column stats are handled as LongColumnStatsData, similarly to
integer types, but high/low value handling is not yet implemented for
timestamps. If for some reason HMS returned high/low values for
timestamps columns a Precondition ("Unsupported type encountered in
setLowAndHighValue()") was hit in Catalogd leading to failing to load
the table.

Impala does not write high/low values for timestamp columns, so I don't
know what led to this state in HMS and could only reproduce the issue
by manipulating TAB_COL_STATS in the backing db of HMS.

Testing:
- only tested manually by manipulating TAB_COL_STATS in HMS's db

Change-Id: If585d2543d49978140dcb7b8d49d6ea50e4a8544
Reviewed-on: http://gerrit.cloudera.org:8080/19548
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>