IMPALA-11965: Fix TestCodegenCache failure when codegen cache disabled by default
The patch fixes the testcase TestCodegenCache failure when the codegen
cache is changed to be disabled by default, because the testcase
assumes the codegen cache is enabled with the default setting.

The solution is to specify a default value to codegen_cache_capacity
in the testcase's start option, so that manually ensures the
codegen cache is on during the test.

Tests:
Passed TestCodegenCache in the exhaustive run with codegen cache
disabled by default.

Change-Id: I749a6ba68553834bdea908741aa7449ed32cd569
Reviewed-on: http://gerrit.cloudera.org:8080/19574
Reviewed-by: Zoltan Borok-Nagy <boroknagyz@cloudera.com>
Reviewed-by: Michael Smith <michael.smith@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>