IMPALA-9551: (Addendum) disable sorting if select list contains struct containing collection
Sorting is not supported if the select list contains collection columns
(see IMPALA-10939). IMPALA-9551 added support for mixed complex types
(collections in structs and structs in collections). However, the case
of having structs containing collections in the select list when sorting
was not handled explicitly. The query

  select id, struct_contains_arr from collection_struct_mix order by id;

resulted in
  ERROR: IllegalStateException: null

After this change, a meaningful error message is given (the same as in
the case of pure collection columns):

  ERROR: IllegalStateException: Sorting is not supported if the select
  list contains collection columns.

The check for collections in the sorting tuple was moved to an earlier
stage of analysis from SingleNodePlanner to QueryStmt, as otherwise we
would hit another precondition check first in the case of structs
containing collections.

Testing:
 - Added tests in mixed-collections-and-structs.test that test sorting
   when a struct in the select list contains an array and a map
   respectively.

Change-Id: I09ac27cba34ee7c6325a7e7895f3a3c9e1a088e5
Reviewed-on: http://gerrit.cloudera.org:8080/19597
Reviewed-by: Csaba Ringhofer <csringhofer@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>