IMPALA-11935: Generate core dumps if ASAN/TSAN/UBSAN built be tests crash
By default, on 64-bit systems, programs built with sanitizers do not
produce minidumps nor coredumps due to the shadow memory - which is used
for house keeping - being huge and the generated dumps would be too
big (~16TB).

https://gcc.gnu.org/bugzilla//show_bug.cgi?id=89868

The shadow memory can be stripped on exit and the core
dumps can be generated by setting the (T|A|UB)SAN_OPTIONS environment
variable before running the program.

https://stackoverflow.com/questions/42851670/how-to-generate-core-dump-on-addresssanitizer-error

This can help investigating crashes in sanitized builds.

Testing:
 - Locally injected an std::abort() to a test case, built it with TSAN
and verified the creation of the core dump and gdb showed the correct
stack frame.

Change-Id: Idd868fe0f666d683084a24808dd0dcd7766b837c
Reviewed-on: http://gerrit.cloudera.org:8080/19598
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>